refactor : 기존 프로젝트의 이미지를 모두 바꾸는 경우에 대한 로직 수정

+++ src/main/java/com/epris/homepage/global/service/FileService.java
@@ -68,6 +68,7 @@ public class FileService {
     /* S3에서 파일 삭제 */
     public void deleteImage(String imageUrl) throws IOException {
         String imageName = getFileNameFromURL(imageUrl);
+        if(imageUrl.equals(null)) throw new CustomException(ErrorCode.INPUT_IS_NULL);
         try {
             if(!amazonS3Client.doesObjectExist(bucket,imageName)) throw new CustomException(ErrorCode.NO_CONTENT_EXIST);
             amazonS3Client.deleteObject(new DeleteObjectRequest(bucket,imageName));
